requirements.txt
webdriver_manager
faker
allure-pytest==2.13.2
pytest==7.4.3
selenium==4.14.0
requests==2.31.0
pytest-xdist==3.3.1
assertpy